implement internal functions
