+++ internal/command/command.go
@@ -40,7 +40,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/onflow/flow-cli/build"
-	"github.com/onflow/flow-cli/internal/services"
 	"github.com/onflow/flow-cli/internal/settings"
 	"github.com/onflow/flow-cli/pkg/flowkit"
 	"github.com/onflow/flow-cli/pkg/flowkit/config"
@@ -53,6 +52,8 @@ import (
 type Run func(
 	args []string,
 	globalFlags GlobalFlags,
+	logger output.Logger,
+	readerWriter flowkit.ReaderWriter,
 	flow flowkit.Services,
 ) (Result, error)
 
@@ -60,16 +61,9 @@ type Run func(
 type RunWithState func(
 	args []string,
 	globalFlags GlobalFlags,
+	logger output.Logger,
 	flow flowkit.Services,
 	state *flowkit.State,
-) (Result, error)
-
-// RunWithInternal runs the command with internal services.
-type RunWithInternal func(
-	args []string,
-	readerWriter flowkit.ReaderWriter,
-	globalFlags GlobalFlags,
-	internal services.CLIServices,
 ) (Result, error)
 
 type Command struct {
@@ -77,7 +71,6 @@ type Command struct {
 	Flags interface{}
 	Run   Run
 	RunS  RunWithState
-	RunI  RunWithInternal
 }
 
 const (
@@ -126,7 +119,6 @@ func (c Command) AddToParent(parent *cobra.Command) {
 
 		// initialize services
 		flow := flowkit.NewFlowkit(state, *network, clientGateway, logger)
-		internal := services.NewInternal(state, clientGateway, logger)
 
 		// skip version check if flag is set
 		if !Flags.SkipVersionCheck {
@@ -147,8 +139,6 @@ func (c Command) AddToParent(parent *cobra.Command) {
 			}
 
 			result, err = c.RunS(args, Flags, flow, state)
-		} else if c.RunI != nil {
-			result, err = c.RunI(args, Flags, internal)
 		} else {
 			panic("command implementation needs to provide run functionality")
 		}
